II Windows 1.3.1.5; II Core; Unit Tests; Misc - Test framework targeting corrected (net46) - Test SaveLoad_Consistency implemented   - Patient.Save and Patient.Load_Process updated     - Now passes missing parameters - Server: when offline, UsageStats stored in AppData cache file   - When online, uploads offline cached UsageStats (#115) - Settings: user settings persist in AppData/config.cfg   - App.Settings saved on Exit(), not on window size changes - Copyright notice date updated

+++ II Core/Classes/Rhythm.Strip.cs
@@ -254,10 +254,10 @@ namespace II.Rhythm {
         private void SetForwardBuffer (Patient patient, bool onClear = false) {
             /* Set the forward edge buffer (a coefficient of lengthSeconds!) to be the length of 2 beats/breaths */
             if (IsCardiac)
-                forwardBuffer = System.Math.Max (1 + (2 * (patient.GetHR_Seconds / Length)),
+                forwardBuffer = System.Math.Max (1 + (2 * ((float)patient.GetHR_Seconds / Length)),
                     (onClear ? 1.1f : forwardBuffer));
             else if (IsRespiratory)
-                forwardBuffer = System.Math.Max (1 + (2 * (patient.GetRR_Seconds / Length)),
+                forwardBuffer = System.Math.Max (1 + (2 * ((float)patient.GetRR_Seconds / Length)),
                     (onClear ? 1.1f : forwardBuffer));
         }
 
@@ -404,11 +404,11 @@ namespace II.Rhythm {
                 p.Cardiac_Rhythm.ECG_Isoelectric (p, this);
             } else if (CanScale) {
                 float fill = (Length * forwardBuffer) - Last (Points).X;
-                Concatenate (Scale (p, Draw.Flat_Line (fill > p.GetHR_Seconds ? fill : p.GetHR_Seconds, 0f)));
+                Concatenate (Scale (p, Draw.Flat_Line (fill > (float)p.GetHR_Seconds ? fill : (float)p.GetHR_Seconds, 0f)));
             } else {
                 /* Fill waveform through to future buffer with flatline */
                 float fill = (Length * forwardBuffer) - Last (Points).X;
-                Concatenate (Draw.Flat_Line (fill > p.GetHR_Seconds ? fill : p.GetHR_Seconds, 0f));
+                Concatenate (Draw.Flat_Line (fill > (float)p.GetHR_Seconds ? fill : (float)p.GetHR_Seconds, 0f));
             }
 
             SortPoints ();
@@ -520,7 +520,7 @@ namespace II.Rhythm {
 
             /* Fill waveform through to future buffer with flatline */
             float fill = (Length * forwardBuffer) - Last (Points).X;
-            Concatenate (Draw.Flat_Line (fill > p.GetRR_Seconds ? fill : p.GetRR_Seconds, 0f));
+            Concatenate (Draw.Flat_Line (fill > (float)p.GetRR_Seconds ? fill : (float)p.GetRR_Seconds, 0f));
 
             SortPoints ();
         }
